refactored the application code

+++ appserver/src/map/chunk.h
@@ -16,10 +16,10 @@
 namespace Chunk {
 
   struct Room {
-    int x;
-    int y;
-    int w;
-    int h;
+    uint8_t x;
+    uint8_t y;
+    uint8_t w;
+    uint8_t h;
   };
 
   struct Map_Chunk {
