+++ appserver/src/objects/components.h
@@ -2,6 +2,7 @@
 //
 // Created by desktop on 9/21/24.
 //
+#include <cstdint>
 #include <functional>
 
 #ifndef BROWSERRPG_COMPONENTS_H
@@ -10,11 +11,11 @@
 #endif // BROWSERRPG_COMPONENTS_H
 
 namespace Component {
-  static const int mapWidth = 99;
+  static const uint8_t mapWidth = 99;
 
   struct Position {
-    int x;
-    int y;
+    int8_t x;
+    int8_t y;
 
     bool operator==(const Position &other) const {
       return x == other.x && y == other.y;
@@ -27,8 +28,8 @@ namespace Component {
     float fGlobalGoal{};
     float fLocalGoal{};
     // position
-    int x{};
-    int y{};
+    int8_t x{};
+    int8_t y{};
     std::vector<sNode *> vecNeighbours;
     sNode *parent{};
   };
