+++ appserver/src/objects/units/spawn.cpp
@@ -27,7 +27,7 @@ namespace Spawn {
     return unitChars[(int)species];
   }
 
-  void Add_Unit(Units::Objects &objects, int level, Component::Position location, int x, int y, const std::string &name, Units::Gender gender, Units::Species species, Units::Class unitClass, Units::Alignment alignment) {
+  void Add_Unit(Units::Objects &objects, uint8_t level, Component::Position location, int8_t x, int8_t y, const std::string &name, Units::Gender gender, Units::Species species, Units::Class unitClass, Units::Alignment alignment) {
     Units::Unit unit(level, location);
     unit.name = name;
     unit.def.gender = gender;
